pyproject.toml
[tool.poetry]
name = "2paq"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = ">=3.8"
spacy = "^3.0.6"
spacy-transformers = "^1.0.3"
ipykernel = "^6.0.1"
jupyter = "^1.0.0"
jsonlines = "^2.0.0"
sentence-transformers = "^2.0.0"
sqlitedict = "^1.7.0"
imdbpy = {git = "https://github.com/alberanid/imdbpy"}
wikipedia = "^1.4.0"
SPARQLWrapper = "^1.8.5"
pytest = "^6.2.5"
number-parser = "^0.2.1"
pendulum = "^2.1.2"
retrying = "^1.3.3"
diskcache = "^5.2.1"
pandas = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
